+++ index.js
@@ -1,3 +1,4 @@
 exports.rules = {
   'no-untracked': require('./rules/no-untracked'),
+  'no-unstaged': require('./rules/no-unstaged'),
 }